pyproject.toml
[project]
name = "office-assistant"
version = "0.1.0"
description = "Office 365 calendar assistant powered by Claude Code and Microsoft Graph"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0",
    "msal>=1.28",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=6.0",
    "respx>=0.21",
    "ruff>=0.8",
    "pre-commit>=4.0",
]

[project.urls]
Repository = "https://github.com/rgilks/office-assistant"
Issues = "https://github.com/rgilks/office-assistant/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/office_assistant"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["office_assistant"]

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 80
omit = [
    "src/office_assistant/__main__.py",
    "src/office_assistant/server.py",
    "src/office_assistant/setup.py",
]

[tool.ruff]
target-version = "py311"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["office_assistant"]
